fix: review comments, moved test to other_test, remoed force_delete_storage entry
