+++ modules/fscloud/variables.tf
@@ -119,9 +119,11 @@ variable "ocp_entitlement" {
 
 variable "force_delete_storage" {
   type        = bool
-  description = "Flag indicating whether or not to delete attached storage when destroying the cluster - Default: false"
-  default     = false
+  description = "Flag indicating whether or not to delete attached storage when destroying the cluster - Default: true"
+  default     = true
+  nullable    = false
 }
+
 variable "existing_cos_id" {
   type        = string
   description = "The COS id of an already existing COS instance"
