+++ variables.tf
@@ -220,8 +220,9 @@ variable "ocp_entitlement" {
 
 variable "force_delete_storage" {
   type        = bool
-  description = "Flag indicating whether or not to delete attached storage when destroying the cluster - Default: false"
-  default     = false
+  description = "Flag indicating whether or not to delete attached storage when destroying the cluster - Default: true"
+  default     = true
+  nullable    = false
 }
 
 variable "cos_name" {
